study_mesh/requirements.txt
streamlit
google-genai
nest-asyncio
python-dotenv
requests
arxiv
plotly
pandas
google-adk
beautifulsoup4